:art: Extract visibleRows to a reselector instead
Only updates if any of the required state input changes
