+++ lib/react/app.js
@@ -4,26 +4,31 @@ import * as actionCreators from '../action-creators'
 import {connect} from 'react-redux'
 import {bindActionCreators} from 'redux'
 import Main from './main'
+import makeVisibleRowsSelector from '../reselectors/visible-rows'
 
-function mapStateToProps (state: State) {
-  return {
-    columnHeaders: state.columnHeaders,
-    initialScanDone: !!state.initialScan.done,
-    initialScanFilesCount: state.initialScan.rawFiles.length,
-    itemsCount: state.sifterResult.total,
-    listHeight: state.config.listHeight,
-    paginationStart: state.pagination.start,
-    query: state.sifterResult.query,
-    rowHeight: state.config.rowHeight,
-    scrollTop: state.scrollTop,
-    sortDirection: state.config.sortDirection,
-    sortField: state.config.sortField,
-    visibleRows: state.visibleRows
+export default function makeApp (columns: Columns) {
+  const visibleRowsSelector = makeVisibleRowsSelector(columns)
+
+  function mapStateToProps (state: State) {
+    return {
+      columnHeaders: state.columnHeaders,
+      initialScanDone: !!state.initialScan.done,
+      initialScanFilesCount: state.initialScan.rawFiles.length,
+      itemsCount: state.sifterResult.total,
+      listHeight: state.config.listHeight,
+      paginationStart: state.pagination.start,
+      query: state.sifterResult.query,
+      rowHeight: state.config.rowHeight,
+      scrollTop: state.scrollTop,
+      sortDirection: state.config.sortDirection,
+      sortField: state.config.sortField,
+      visibleRows: visibleRowsSelector(state)
+    }
   }
-}
 
-function mapDispatchToProps (dispatch: Dispatch) {
-  return {actions: bindActionCreators(actionCreators, dispatch)}
-}
+  function mapDispatchToProps (dispatch: Dispatch) {
+    return {actions: bindActionCreators(actionCreators, dispatch)}
+  }
 
-export default connect(mapStateToProps, mapDispatchToProps)(Main)
+  return connect(mapStateToProps, mapDispatchToProps)(Main)
+}
