+++ lib/reducers/index.js
@@ -8,14 +8,12 @@ import scrollTop from './scroll-top'
 import selectedNote from './selected-note'
 import setupNotes from './notes'
 import setupSifterResult from './sifter-result'
-import setupVisibleRows from './visible-rows'
 
 export default function setupRoot (columns: Columns, notesFields: NotesFields) {
   const columnHeaders = setupColumnHeaders(columns)
   const config = setupConfig()
   const notes = setupNotes(notesFields)
   const sifterResult = setupSifterResult(notesFields)
-  const visibleRows = setupVisibleRows(columns)
 
   return function root (state: State, action: Action) {
     if (!state) {
@@ -33,15 +31,12 @@ export default function setupRoot (columns: Columns, notesFields: NotesFields) {
       pagination: state.pagination,
       scrollTop: state.scrollTop,
       selectedNote: selectedNote(state.selectedNote, action),
-      sifterResult: state.sifterResult,
-      visibleRows: state.visibleRows
+      sifterResult: state.sifterResult
     }
 
     nextState.sifterResult = sifterResult(state.sifterResult, action, nextConfig, nextState.notes)
     nextState.scrollTop = scrollTop(state.scrollTop, action, nextConfig, nextState.selectedNote)
     nextState.pagination = pagination(state.pagination, action, nextConfig, nextState.scrollTop)
-
-    nextState.visibleRows = visibleRows(state.visibleRows, action, nextState)
 
     return nextState
   }
