+++ lib/session.js
@@ -9,7 +9,7 @@ import Disposables from './disposables'
 import {startInitialScan, dispose} from './action-creators.js'
 import epicMiddleware from './epics'
 import setupRootReducer from './reducers/index'
-import App from './react/app'
+import makeApp from './react/app'
 import TogglePanel from './toggle-panel'
 import ToggleAtomWindow from './toggle-atom-window'
 import RestartSessionForNewConfigToTakeEffect from './restart-session-for-new-config-to-take-effect'
@@ -27,6 +27,7 @@ export default class Session {
     const middlewares = applyMiddleware(epicMiddleware)
     const store = createStore(rootReducer, middlewares)
 
+    const App = makeApp(columns)
     render(<Provider store={store}>
       <App />
     </Provider>, panel.getItem())
