+++ package.json
@@ -47,6 +47,7 @@
     "react-redux": "5.0.1",
     "redux": "3.6.0",
     "redux-observable": "0.12.2",
+    "reselect": "2.5.4",
     "rxjs": "5.0.2",
     "sifter": "0.5.2"
   },
